ENH: Install specific TotalSegmentator Python package version
Until now, there was an option to install latest TotalSegmentator Python package from GitHub.
However, since the download URL was always the same, pip always used the same existing package in the cache (did not retrieve the latest content of the URL).
There were also issues with options and behavior changing significantly between TotalSegmentator Python package versions, so updating the package separately from the Slicer module appeared to be risky.

Therefore, behavior of the module is now changed so that the required TotalSegmentator Python package version (defined by download URL) is specified in the module logic constructor. When segmentation is started then the package version is checked and the package is updated if needed.

An option for forced reinstallation was kept for troubleshooting and restoration of corrupted installations.